+++ rio_viz/app.py
@@ -2,15 +2,13 @@
 
 import pathlib
 import urllib.parse
-from enum import Enum
 from functools import partial
 from typing import Any, Dict, Optional, Type, Union
 
 import attr
 import rasterio
 import uvicorn
-from fastapi import FastAPI, Query
-from rasterio.enums import Resampling
+from fastapi import APIRouter, Depends, FastAPI, Query
 from starlette.concurrency import run_in_threadpool
 from starlette.middleware.cors import CORSMiddleware
 from starlette.middleware.gzip import GZipMiddleware
@@ -18,13 +16,14 @@ from starlette.requests import Request
 from starlette.responses import HTMLResponse
 from starlette.templating import Jinja2Templates
 
-from rio_tiler.colormap import cmap
 from rio_tiler.io import AsyncBaseReader
 from rio_tiler.models import Info, Metadata
 
+from .dependencies import ImageParams
 from .models.mapbox import TileJSON
-from .ressources.enums import ImageType, MimeTypes, VectorType
-from .ressources.responses import TileResponse, XMLResponse
+from .ressources.enums import ImageType, TileType
+from .ressources.responses import ImageResponse, XMLResponse
+from .utils import Timer
 
 try:
     from rio_tiler_mvt.mvt import encoder as mvt_encoder  # noqa
@@ -36,16 +35,7 @@ except ModuleNotFoundError:
 
 template_dir = str(pathlib.Path(__file__).parent.joinpath("templates"))
 templates = Jinja2Templates(directory=template_dir)
-
-
 _mvt_encoder = partial(run_in_threadpool, mvt_encoder)
-
-ResamplingNames = Enum(  # type: ignore
-    "ResamplingNames", [(r.name, r.name) for r in Resampling]
-)
-ColorMapNames = Enum(  # type: ignore
-    "ColorMapNames", [(a, a) for a in sorted(cmap.list())]
-)
 
 
 @attr.s
@@ -69,10 +59,15 @@ class viz:
     layers: Optional[str] = attr.ib(default=None)
     nodata: Optional[Union[str, int, float]] = attr.ib(default=None)
 
+    router: Optional[APIRouter] = attr.ib(init=False)
+
     def __attrs_post_init__(self):
         """Update App."""
+        self.router = APIRouter()
+
         self.register_middleware()
         self.register_routes()
+        self.app.include_router(self.router)
 
     def register_middleware(self):
         """Register Middleware to the FastAPI app."""
@@ -113,31 +108,42 @@ class viz:
     def register_routes(self):
         """Register routes to the FastAPI app."""
 
-        @self.app.get(
+        @self.router.get(
             r"/tiles/{z}/{x}/{y}",
             responses={
                 200: {
                     "content": {
                         "image/png": {},
-                        "image/jpg": {},
+                        "image/jpeg": {},
                         "image/webp": {},
+                        "image/jp2": {},
+                        "image/tiff; application=geotiff": {},
+                        "application/x-binary": {},
                         "application/x-protobuf": {},
                     },
                     "description": "Return a tile.",
                 }
             },
-            response_class=TileResponse,
+            response_class=ImageResponse,
             description="Read COG and return a tile",
         )
-        @self.app.get(
+        @self.router.get(
             r"/tiles/{z}/{x}/{y}.{format}",
             responses={
                 200: {
-                    "content": {"image/png": {}, "image/jpg": {}, "image/webp": {}},
+                    "content": {
+                        "image/png": {},
+                        "image/jpeg": {},
+                        "image/webp": {},
+                        "image/jp2": {},
+                        "image/tiff; application=geotiff": {},
+                        "application/x-binary": {},
+                        "application/x-protobuf": {},
+                    },
                     "description": "Return an image.",
                 }
             },
-            response_class=TileResponse,
+            response_class=ImageResponse,
             description="Read COG and return a tile",
         )
         async def tile(
@@ -145,20 +151,11 @@ class viz:
             x: int,
             y: int,
             scale: int = Query(2, gt=0, lt=4),
-            format: Optional[Union[ImageType, VectorType]] = None,
+            format: Optional[TileType] = None,
             indexes: Optional[str] = Query(
                 None, description="Coma (',') delimited band indexes"
             ),
-            rescale: Optional[str] = Query(
-                None, description="Coma (',') delimited Min,Max bounds"
-            ),
-            color_formula: Optional[str] = Query(None, description="rio-color formula"),
-            color_map: Optional[ColorMapNames] = Query(
-                None, description="rio-tiler's colormap name"
-            ),
-            resampling_method: ResamplingNames = Query(
-                ResamplingNames.nearest, description="Resampling method."  # type: ignore
-            ),
+            image_params: ImageParams = Depends(),
             feature_type: str = Query(
                 None,
                 title="Feature type (Only for Vector)",
@@ -166,56 +163,238 @@ class viz:
             ),
         ):
             """Handle /tiles requests."""
+            timings = []
+            headers: Dict[str, str] = {}
+
             tilesize = scale * 256
 
-            if format and format in [VectorType.pbf, VectorType.mvt]:
+            if format and format in [TileType.pbf, TileType.mvt]:
                 tilesize = 128
 
-            async with self.reader(self.src_path) as src_dst:  # type: ignore
-                kwargs = self._get_options(src_dst, indexes)
-                tile_data = await src_dst.tile(
-                    x,
-                    y,
-                    z,
-                    tilesize=tilesize,
-                    resampling_method=resampling_method.name,
-                    **kwargs,
-                )
+            with Timer() as t:
+                async with self.reader(self.src_path) as src_dst:  # type: ignore
+                    kwargs = self._get_options(src_dst, indexes)
+                    tile_data = await src_dst.tile(
+                        x,
+                        y,
+                        z,
+                        tilesize=tilesize,
+                        resampling_method=image_params.resampling_method.name,
+                        **kwargs,
+                    )
+                    colormap = image_params.colormap or getattr(
+                        src_dst, "colormap", None
+                    )
+                    bands = kwargs.get("bands", kwargs.get("assets", None)) or [
+                        f"{ix + 1}" for ix in range(tile_data.count)
+                    ]
+            timings.append(("dataread", round(t.elapsed * 1000, 2)))
 
-                bands = kwargs.get("bands", kwargs.get("assets", None)) or [
-                    f"{ix + 1}" for ix in range(tile_data.count)
-                ]
-
-            if format and format in [VectorType.pbf, VectorType.mvt]:
-                content = await _mvt_encoder(
-                    tile_data.data, tile_data.mask, bands, feature_type=feature_type,
-                )  # type: ignore
+            if format and format in [TileType.pbf, TileType.mvt]:
+                with Timer() as t:
+                    content = await _mvt_encoder(
+                        tile_data.data,
+                        tile_data.mask,
+                        bands,
+                        feature_type=feature_type,
+                    )  # type: ignore
+                timings.append(("format", round(t.elapsed * 1000, 2)))
             else:
-                in_range = list(map(float, rescale.split(","))) if rescale else None
-                image = tile_data.post_process(
-                    in_range=in_range, color_formula=color_formula
-                )
-
                 if not format:
-                    format = ImageType.jpg if tile_data.mask.all() else ImageType.png
+                    format = TileType.jpeg if tile_data.mask.all() else TileType.png
 
-                colormap = (
-                    cmap.get(color_map.value)
-                    if color_map
-                    else getattr(src_dst, "colormap", None)
+                with Timer() as t:
+                    image = tile_data.post_process(
+                        in_range=image_params.rescale_range,
+                        color_formula=image_params.color_formula,
+                    )
+                timings.append(("postprocess", round(t.elapsed * 1000, 2)))
+
+                with Timer() as t:
+                    content = image.render(
+                        img_format=format.driver, colormap=colormap, **format.profile,
+                    )
+                timings.append(("format", round(t.from_start * 1000, 2)))
+
+            headers["Server-Timing"] = ", ".join(
+                [f"{name};dur={time}" for (name, time) in timings]
+            )
+            return ImageResponse(content, media_type=format.mimetype, headers=headers)
+
+        @self.router.get(
+            r"/preview",
+            responses={
+                200: {
+                    "content": {
+                        "image/png": {},
+                        "image/jpeg": {},
+                        "image/webp": {},
+                        "image/jp2": {},
+                        "image/tiff; application=geotiff": {},
+                        "application/x-binary": {},
+                    },
+                    "description": "Return a preview.",
+                }
+            },
+            response_class=ImageResponse,
+            description="Return a preview",
+        )
+        @self.router.get(
+            r"/preview.{format}",
+            responses={
+                200: {
+                    "content": {
+                        "image/png": {},
+                        "image/jpeg": {},
+                        "image/webp": {},
+                        "image/jp2": {},
+                        "image/tiff; application=geotiff": {},
+                        "application/x-binary": {},
+                    },
+                    "description": "Return a preview.",
+                }
+            },
+            response_class=ImageResponse,
+            description="Return a preview.",
+        )
+        async def preview(
+            format: Optional[ImageType] = None,
+            indexes: Optional[str] = Query(
+                None, description="Coma (',') delimited band indexes"
+            ),
+            max_size: int = 1024,
+            image_params: ImageParams = Depends(),
+        ):
+            """Handle /preview requests."""
+            timings = []
+            headers: Dict[str, str] = {}
+
+            with Timer() as t:
+                async with self.reader(self.src_path) as src_dst:  # type: ignore
+                    kwargs = self._get_options(src_dst, indexes)
+                    data = await src_dst.preview(
+                        max_size=max_size,
+                        resampling_method=image_params.resampling_method.name,
+                        **kwargs,
+                    )
+                    colormap = image_params.colormap or getattr(
+                        src_dst, "colormap", None
+                    )
+            timings.append(("dataread", round(t.elapsed * 1000, 2)))
+
+            if not format:
+                format = ImageType.jpeg if data.mask.all() else ImageType.png
+
+            with Timer() as t:
+                image = data.post_process(
+                    in_range=image_params.rescale_range,
+                    color_formula=image_params.color_formula,
                 )
+            timings.append(("postprocess", round(t.elapsed * 1000, 2)))
+
+            with Timer() as t:
                 content = image.render(
                     img_format=format.driver, colormap=colormap, **format.profile,
                 )
+            timings.append(("format", round(t.from_start * 1000, 2)))
 
-            return TileResponse(content, media_type=MimeTypes[format.value].value)
+            headers["Server-Timing"] = ", ".join(
+                [f"{name};dur={time}" for (name, time) in timings]
+            )
+            return ImageResponse(content, media_type=format.mimetype, headers=headers)
 
-        @self.app.get(
+        @self.router.get(
+            r"/part",
+            responses={
+                200: {
+                    "content": {
+                        "image/png": {},
+                        "image/jpeg": {},
+                        "image/webp": {},
+                        "image/jp2": {},
+                        "image/tiff; application=geotiff": {},
+                        "application/x-binary": {},
+                    },
+                    "description": "Return a part of a dataset.",
+                }
+            },
+            response_class=ImageResponse,
+            description="Return a preview.",
+        )
+        @self.router.get(
+            r"/part.{format}",
+            responses={
+                200: {
+                    "content": {
+                        "image/png": {},
+                        "image/jpeg": {},
+                        "image/webp": {},
+                        "image/jp2": {},
+                        "image/tiff; application=geotiff": {},
+                        "application/x-binary": {},
+                    },
+                    "description": "Return a part of a dataset.",
+                }
+            },
+            response_class=ImageResponse,
+            description="Return a preview.",
+        )
+        async def part(
+            format: Optional[ImageType] = Query(None, description="Output image type."),
+            bbox: str = Query(
+                ..., description="Bounding box in form of 'minx,miny,maxx,maxy'"
+            ),
+            indexes: Optional[str] = Query(
+                None, description="Coma (',') delimited band indexes"
+            ),
+            max_size: int = 1024,
+            image_params: ImageParams = Depends(),
+        ):
+            """Handle /part requests."""
+            timings = []
+            headers: Dict[str, str] = {}
+
+            with Timer() as t:
+                async with self.reader(self.src_path) as src_dst:  # type: ignore
+                    kwargs = self._get_options(src_dst, indexes)
+                    data = await src_dst.part(
+                        list(map(float, bbox.split(","))),
+                        max_size=max_size,
+                        resampling_method=image_params.resampling_method.name,
+                        **kwargs,
+                    )
+                    colormap = image_params.colormap or getattr(
+                        src_dst, "colormap", None
+                    )
+            timings.append(("dataread", round(t.elapsed * 1000, 2)))
+
+            if not format:
+                format = ImageType.jpeg if data.mask.all() else ImageType.png
+
+            with Timer() as t:
+                image = data.post_process(
+                    in_range=image_params.rescale_range,
+                    color_formula=image_params.color_formula,
+                )
+            timings.append(("postprocess", round(t.elapsed * 1000, 2)))
+
+            with Timer() as t:
+                content = image.render(
+                    img_format=format.driver, colormap=colormap, **format.profile,
+                )
+            timings.append(("format", round(t.from_start * 1000, 2)))
+
+            headers["Server-Timing"] = ", ".join(
+                [f"{name};dur={time}" for (name, time) in timings]
+            )
+            return ImageResponse(content, media_type=format.mimetype, headers=headers)
+
+        @self.router.get(
             "/point", responses={200: {"description": "Return a point value."}},
         )
         async def point(
             coordinates: str = Query(
-                ..., title="Coma (',') delimited lon,lat coordinates"
+                ..., description="Coma (',') delimited lon,lat coordinates"
             ),
         ):
             """Handle /point requests."""
@@ -226,7 +405,7 @@ class viz:
 
             return {"coordinates": [lon, lat], "value": results}
 
-        @self.app.get(
+        @self.router.get(
             "/metadata",
             response_model=Union[Dict[str, Metadata], Metadata],
             response_model_exclude={"minzoom", "maxzoom", "center"},
@@ -236,6 +415,7 @@ class viz:
         async def metadata(
             pmin: float = 2.0,
             pmax: float = 98.0,
+            max_size: int = 1024,
             indexes: Optional[str] = Query(
                 None, title="Coma (',') delimited band indexes"
             ),
@@ -243,9 +423,9 @@ class viz:
             """Handle /metadata requests."""
             async with self.reader(self.src_path) as src_dst:  # type: ignore
                 kwargs = self._get_options(src_dst, indexes)
-                return await src_dst.metadata(pmin, pmax, **kwargs)
+                return await src_dst.metadata(pmin, pmax, max_size=max_size, **kwargs)
 
-        @self.app.get(
+        @self.router.get(
             "/info",
             response_model=Union[Dict[str, Info], Info],
             response_model_exclude={"minzoom", "maxzoom", "center"},
@@ -258,7 +438,7 @@ class viz:
                 kwargs = self._get_options(src_dst)
                 return await src_dst.info(**kwargs)
 
-        @self.app.get(
+        @self.router.get(
             "/tilejson.json",
             response_model=TileJSON,
             responses={200: {"description": "Return a tilejson"}},
@@ -266,22 +446,11 @@ class viz:
         )
         async def tilejson(
             request: Request,
-            tile_format: Optional[Union[ImageType, VectorType]] = None,
+            tile_format: Optional[TileType] = None,
             indexes: Optional[str] = Query(  # noqa
                 None, description="Coma (',') delimited band indexes"
             ),
-            rescale: Optional[str] = Query(  # noqa
-                None, description="Coma (',') delimited Min,Max bounds"
-            ),
-            color_formula: Optional[str] = Query(
-                None, description="rio-color formula"
-            ),  # noqa
-            color_map: Optional[ColorMapNames] = Query(  # noqa
-                None, description="rio-tiler's colormap name"
-            ),
-            resampling_method: ResamplingNames = Query(  # noqa
-                ResamplingNames.nearest, description="Resampling method."  # type: ignore
-            ),
+            image_params: ImageParams = Depends(),  # noqa
             feature_type: str = Query(  # noqa
                 None, title="Feature type", regex="^(point)|(polygon)$"
             ),
@@ -315,7 +484,7 @@ class viz:
                 tiles=[tile_url],
             )
 
-        @self.app.get(
+        @self.router.get(
             "/index.html",
             responses={200: {"description": "Simple COG viewer."}},
             response_class=HTMLResponse,
@@ -336,7 +505,7 @@ class viz:
                 media_type="text/html",
             )
 
-        @self.app.get("/WMTSCapabilities.xml", response_class=XMLResponse)
+        @self.router.get("/WMTSCapabilities.xml", response_class=XMLResponse)
         async def wmts(
             request: Request,
             tile_format: ImageType = Query(
@@ -345,18 +514,7 @@ class viz:
             indexes: Optional[str] = Query(  # noqa
                 None, description="Coma (',') delimited band indexes"
             ),
-            rescale: Optional[str] = Query(  # noqa
-                None, description="Coma (',') delimited Min,Max bounds"
-            ),
-            color_formula: Optional[str] = Query(
-                None, description="rio-color formula"
-            ),  # noqa
-            color_map: Optional[ColorMapNames] = Query(  # noqa
-                None, description="rio-tiler's colormap name"
-            ),
-            resampling_method: ResamplingNames = Query(  # noqa
-                ResamplingNames.nearest, description="Resampling method."  # type: ignore
-            ),
+            image_params: ImageParams = Depends(),  # noqa
             feature_type: str = Query(  # noqa
                 None, title="Feature type", regex="^(point)|(polygon)$"
             ),
@@ -404,8 +562,6 @@ class viz:
                 </TileMatrix>"""
                 tileMatrix.append(tm)
 
-            media_type = MimeTypes[tile_format.value].value
-
             return templates.TemplateResponse(
                 "wmts.xml",
                 {
@@ -415,7 +571,7 @@ class viz:
                     "tileMatrix": tileMatrix,
                     "title": "Cloud Optimized GeoTIFF",
                     "layer_name": "cogeo",
-                    "media_type": media_type,
+                    "media_type": tile_format.mimetype,
                 },
                 media_type="application/xml",
             )
@@ -429,6 +585,11 @@ class viz:
     def template_url(self) -> str:
         """Get simple app template url."""
         return f"http://{self.host}:{self.port}/index.html"
+
+    @property
+    def docs_url(self) -> str:
+        """Get simple app template url."""
+        return f"http://{self.host}:{self.port}/docs"
 
     def start(self):
         """Start tile server."""